refactor: car wall clean up works
