+++ src/environments/environments/CarWallReset.py
@@ -12,7 +12,7 @@ from geometry_msgs.msg import Pose, Point
 
 from ament_index_python import get_package_share_directory
 
-from .util.util import get_quaternion_from_euler
+from .util import get_quaternion_from_euler
 
 class CarWallReset(Node):
     def __init__(self):
