--- src/environments/environments/util/__init__.py
@@ -1,1 +1,0 @@
-import util
