+++ src/reinforcement_learning/config/train.yaml
@@ -1,6 +1,6 @@
 train:
   ros__parameters:
-    environment: 'CarTrack2'
+    environment: 'CarWall'
     max_steps_exploration: 5000
     max_steps_training: 1000000
     reward_range: 1.0
